ranker/requirements-freeze.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python_protos @ file:///Users/raphael/src/python-grpc-toy/python_protos/runtime
python_protos-stubs @ file:///Users/raphael/src/python-grpc-toy/python_protos/stubs
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.2
